fix: query streaming reliability issues
Use a result buffer to ensure no results are dropped. Use a promise that resolves with true when an "end" event is received or resolves with false when a "result" event is received. Be careful to ensure that the promise is reset immediately after it's resolved, so an "end" event is never missed.