feat: support static `INSERT` statements
Your SQL files may contain INSERT statements (i.e. to populate a lookup table). Each "static row" is individually tracked in the `nano.inserts` table, so pg-nano can update/delete the row if you update/remove it in your SQL file.

In order to support this, we no longer rely on pg-schema-diff to execute the migration plan, but we still use it to generate the migration plan. We may drop pg-schema-diff altogether in the future, since it repeats a lot of work already done by pg-nano.

Various events have been added/renamed in this commit.